udpate deps and move code into new proj
