+++ build.sbt
@@ -1,4 +1,4 @@
-scalaVersion := "2.11.8"
+scalaVersion in ThisBuild := "2.11.8"
 
 name := "media-atom-maker"
 
@@ -52,9 +52,10 @@ lazy val appDistSettings = Seq(
 lazy val atomPublisher = (project in file("./atom-publisher-lib"))
 
 lazy val atomManagerPlay = (project in file("./atom-manager-play-lib"))
+  .dependsOn(atomPublisher)
 
 lazy val root = (project in file("."))
   .enablePlugins(PlayScala, RiffRaffArtifact, UniversalPlugin)
   .settings(appDistSettings)
-  .dependsOn(atomPublisher)
+  .dependsOn(atomManagerPlay)
