+++ test/ApiSpec.scala
@@ -2,7 +2,7 @@ package test
 
 import cats.data.Xor
 import com.gu.atom.publish.AtomPublisher
-import data.{ DataStore, VersionConflictError }
+import com.gu.atom.data.{ DataStore, VersionConflictError }
 import org.scalatest.mock.MockitoSugar
 import org.mockito.Mockito._
 import org.mockito.Matchers._
